Fixed string related memory leak.
On hash_map_insert(), when both keys are identical the pointer that keeps track of the entry key is reassigned to the newly allocated string, leaking memory by losing the pointer to the old key string.
With this fix, when we have identical keys, we free the new one and keep the old.